Split index into app.js and index.js
