+++ server/tests/prizes.test.js
@@ -1,0 +1,1 @@
+
